refactor code and bring logic to utils functions
